Better PRNG function for Zobrist keys

+++ includes/Computer.h
@@ -15,7 +15,7 @@ struct PositionSave
 {
 	uint64_t hash;
 	int8_t depth;
-	float score;
+	int score;
 	bool ancient;
 	NodeType type;
 	Move best_move;
@@ -30,7 +30,7 @@ public:
 
 	Computer();
 
-	std::pair<Move, float> find_move(Board board, uint8_t depth, float alpha, float beta, bool maximize);
+	std::pair<Move, int> find_move(Board board, uint8_t depth, int alpha, int beta, bool maximize);
 	void move(Board& board, uint8_t depth);
 	uint64_t signature_hash(const Board& board) const;
 	uint64_t get_key(uint8_t x, uint8_t y, Type type) const;
